use Integer as the voltage data type instead of Double

+++ src/main/java/com/emilburzo/ambientlogger/model/Reading.java
@@ -5,7 +5,7 @@ public class Reading {
     public Integer id;
     public Double temperature;
     public Double humidity;
-    public Double voltage;
+    public Integer voltage;
     public Long timestamp;
 
 }
